Issue the group kill only before Wait, and build the probe test on unix alone
Installing the group kill as cmd.Cancel let the exec package's watcher
fire between Process.Wait reaping the leader and Wait synchronizing with
it — after the group ID could have been recycled, the race this change
exists to exclude. The kill now happens in one place, on this goroutine,
strictly before Wait; a leader still running past that point is killed
alone by the default cancel. The grandchild assertion uses syscall.Kill,
so the test moves behind a unix build tag instead of a runtime skip.
